requirements.txt
psycopg2
django-cas-ng
